+++ Sources/RAW_macros/RAW_staticbuff_macro.swift
@@ -224,7 +224,25 @@ public struct RAW_staticbuff_macro:MemberMacro, ExtensionMacro {
 						context.addDiagnostics(from:StoredVariablesUnsupported(), node:curVar)
 					}
 				}
-				
+
+				declString.append(DeclSyntax("""
+					/// initialize the static buffer from  its raw representation store type. behavior is undefined if the raw representation is shorter than the assumed size of the static buffer.
+					\(asStruct.modifiers) init(RAW_staticbuff storetype:consuming RAW_staticbuff_storetype) {
+						#if DEBUG
+						assert(MemoryLayout<Self>.size == MemoryLayout<RAW_staticbuff_storetype>.size, "static buffer type size mismatch. this is a misuse of the macro")
+						assert(MemoryLayout<Self>.stride == MemoryLayout<RAW_staticbuff_storetype>.stride, "static buffer type stride mismatch. this is a misuse of the macro")
+						assert(MemoryLayout<Self>.alignment == MemoryLayout<RAW_staticbuff_storetype>.alignment, "static buffer type alignment mismatch. this is a misuse of the macro")
+						#endif
+						\(varName) = storetype
+					}
+				"""))
+
+				declString.append(DeclSyntax("""
+					/// borrow the raw representation of the static buffer.
+					\(asStruct.modifiers) consuming func RAW_staticbuff() -> RAW_staticbuff_storetype {
+						return \(varName)
+					}
+				"""))
 			// emit a comparison operator that equates to a linear execution of the various members of the staticbuff
 			case .concatType(let tokens):
 				// assemble the primary extension declaration.
@@ -361,6 +379,27 @@ public struct RAW_staticbuff_macro:MemberMacro, ExtensionMacro {
 						\(raw:buildCompare.joined(separator: "\n"))
 
 						return compare_result
+					}
+				"""))
+
+				declString.append(DeclSyntax("""
+					\(raw:asStruct.modifiers) init(RAW_staticbuff storetype:consuming RAW_staticbuff_storetype) {
+						#if DEBUG
+						assert(MemoryLayout<Self>.size == MemoryLayout<RAW_staticbuff_storetype>.size, "static buffer type size mismatch. this is a misuse of the macro")
+						assert(MemoryLayout<Self>.stride == MemoryLayout<RAW_staticbuff_storetype>.stride, "static buffer type stride mismatch. this is a misuse of the macro")
+						assert(MemoryLayout<Self>.alignment == MemoryLayout<RAW_staticbuff_storetype>.alignment, "static buffer type alignment mismatch. this is a misuse of the macro")
+						#endif
+						self = withUnsafePointer(to:&storetype) { ptr in
+							return UnsafeRawPointer(ptr).load(as:Self.self)
+						}
+					}
+				"""))
+
+				declString.append(DeclSyntax("""
+					\(raw:asStruct.modifiers) consuming func RAW_staticbuff() -> RAW_staticbuff_storetype {
+						return withUnsafePointer(to:&self) { ptr in
+							return UnsafeRawPointer(ptr).load(as:RAW_staticbuff_storetype.self)
+						}
 					}
 				"""))
 			}